pjt_final/server/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.46.0
django-cors-headers==3.10.0
django-extensions==3.1.5
django-rest-auth==0.9.5
django-seed==0.3.1
djangorestframework==3.12.4
Faker==9.8.3
idna==3.3
ipython==7.29.0
jedi==0.18.1
matplotlib-inline==0.1.3
oauthlib==3.1.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.22
psycopg2==2.9.2
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
